CESMII edge-to-database
This is the code for sending vibration data from a MEMS accelerometer ADXL-345 to a URL which has a database set-up for data ingestion. The sampling time and number of axes to be sampled can be changed in parameters or functions. Link to the AdaFruit ADXL-345 python library is: https://github.com/adafruit/Adafruit_Python_ADXL345.